create card from form

+++ recipe-client/src/components/pages/SavedCardsPage.jsx
@@ -1,33 +1,38 @@
 import { useState } from 'react';
-import { json } from 'react-router-dom';
-
+// import { json } from 'react-router-dom';
+import CreateCard from '../elements/createCard';
 
 function SavedCardsPage({ x, setData, }) {
-  const [input, setInput] = useState("inti")
+
   const [editState, setEditState] = useState(false);
   const [update, setUpdate] = useState('');
   const [currentCard, setCurrentCard] = useState();
 
-  const handleInput = (e) => {
-    setInput(e.target.value);
+  const handleShow = (e) => {
+    const x = (e) => {
+      console.log(e.target.id)// 
+      setCurrentCard(e.target.id)
 
-  }
+
+    };
+
+    currentCard === e.target.id ? setCurrentCard(null) : x(e);
+
+
+  };
+
   const handleUpdate = (e) => {
     setUpdate(e.target.value);
 
 
   }
-  const handleSubmit = (e) => {
-    e.preventDefault();
 
-    create();
-    setInput("");
-  }
 
   const handle = (e) => {
-    setEditState(!editState);
     currentCard === e.target.id ? setCurrentCard(null) : null
     if (e.target.innerHTML === "edit" && !editState) {
+      setEditState(!editState);
+
       setCurrentCard(e.target.id);
       setUpdate('')
       // setEditState(!editState);
@@ -35,30 +40,32 @@ function SavedCardsPage({ x, setData, }) {
       if (update === '') {
         setUpdate(() => {
 
-          const title = x.filter((r) => {
+          const label = x.filter((r) => {
             return r._id === e.target.id
           })
-          console.log('updatedTItle', title[0].title);
+          console.log('updatedTItle', label[0].label);
 
-          return title[0].title
+          return label[0].label
         });
 
 
       }
 
-      console.log("before IF", editState)
+
 
 
 
     } else if (e.target.innerHTML === "submit" && editState) {
-      if (update !== '') {
+      setEditState(!editState);
+
+      if (update !== '' && update.length > 0 && e.target.id === currentCard) {
         console.log("IF", editState)
         setData((prevState) => {
           console.log('update', update);
 
           const updatedData = prevState.map((r, i) => {
             if (r._id === e.target.id) {
-              r.title = update;
+              r.label = update;
             }
             return r;
           })
@@ -74,7 +81,7 @@ function SavedCardsPage({ x, setData, }) {
 
         fetch(`http://localhost:5000/api/updateCard/${e.target.id}`, {
           method: "PATCH",
-          body: JSON.stringify({ title: update }),
+          body: JSON.stringify({ label: update }),
           headers: {
             "Content-Type": "application/json",
           }
@@ -87,7 +94,7 @@ function SavedCardsPage({ x, setData, }) {
 
 
     }
-    else {
+    else if (e.target.innerHTML === "Delete") {
       // delete
       fetch(`http://localhost:5000/api/deleteCard/${e.target.id}`, {
         method: 'DELETE',
@@ -99,59 +106,53 @@ function SavedCardsPage({ x, setData, }) {
     }
 
   }
-  const create = async () => {
 
-    if (input.length > 0) {
-      await fetch(" http://localhost:5000/api/", {
-        method: "POST",
-        body: JSON.stringify({ title: input }),
-        headers: {
-          "Content-Type": "application/json",
-        }
-      }).then(res => res.json()).then(updatedata => { setData([...x, updatedata]) })
-    } else {
 
-    }
 
-  }
 
   return (
-    <div className="h-screen flex flex-col items-center  mt-14 ">
+    <div className=" flex flex-col items-center  mt-14 min-h-[100vh]">
 
 
-      <form className="flex flex-col" method="post" required>
-        <input placeholder="name" value={input} onChange={handleInput} type='text' required name="newCard" id="" />
-        <input placeholder='type of cusine ' />
-        <input placeholder='ingredients' />
-        <button className="bg-black " type="submit" onClick={handleSubmit}>Create New Recipie</button>
-      </form>
-
-      <div className='flex flex-wrap justify-between p-8  '>
+      <CreateCard x={x} setData={setData} />
+      <div className='px-4 mt-14 cardContainer flex flex-col sm:flex-row sm:flex-wrap  items-center sm:items-start max-w-[1100px] m-auto  '>
         {x ? x.map((recipe, i) => {
-
           return (
-            <div className="bg-white text-black" id={recipe._id} >
-              {<p>{recipe.title}</p>}
-              {currentCard === recipe._id && <input id={recipe._id} type="text" key={i} className=" w-40 border flex justify-between" value={
-                update
-              } onChange={handleUpdate} />}
-              <span className={`  p-4  `}>
-                <h4>{'pasta'}</h4>
+            <div className={` bg-white min-w-[191.391px] max-w-[30%] text-black  mb-5 mr-5 flex  flex-col w-1/2  justify-between${currentCard === recipe._id ? "bg-red-500 h-fit " : "min-h-[225px] max-h-[250px]"
+              }
+        `} id={recipe._id} >
 
-                <p>Dish type: {"dinner"}</p>
+              {currentCard === recipe._id && editState === true ? <input id={recipe._id} type="text" key={i} className=" w-40 border flex justify-between" value={
+                update
+              } onChange={handleUpdate} /> : null}
+
+              <span className={`  p-4  `}>
+                <h4>{recipe.label}</h4>
+
+                <p>Dish type: {recipe.cusineType}</p>
                 <p>ingredients</p>
                 <hr />
+                {currentCard === recipe._id ?
+                  <ul>
+                    <li>hello</li>
+                    {recipe.ingredients.map((ing, i) => {
 
-                <ul>
-                  <li >pebbles</li>
-                  <li>rocks</li>
-                  <li>Candy and stuff</li>
+                      return (<li>{ing}</li>)
+                    })}
 
-                </ul>
+                  </ul> : null
+                }
 
               </span>
-              <button id={recipe._id} className="bg-blue-700" type='submit' onClick={handle}>{currentCard === recipe._id ? "submit" : "edit"}</button ><button id={recipe._id} className="bg-red-700" type='submit' onClick={handle}>Delete</button >
+              <button id={recipe._id} className="bg-blue-700" type='submit' onClick={handle}>{currentCard === recipe._id && editState ? "submit" : "edit"}
+              </button >
+              <button id={recipe._id} className="bg-red-700" type='submit' onClick={handle}>Delete</button >
 
+
+              <button id={recipe._id} onClick={(e) => {
+                console.log(e.target.id)
+                handleShow(e)
+              }} className="bg-blue-300 " >{currentCard === recipe._id ? "hide ing" : "show ing"}</button>
             </div>
 
           )
